LDE-714: moved tests for delete_cube
